Remove oauth tokens
