+++ src/vivarium_cluster_tools/psimulate/utilities.py
@@ -1,6 +1,4 @@
-from bdb import BdbQuit
 from datetime import datetime
-import functools
 import math
 import os
 from pathlib import Path
@@ -141,28 +139,6 @@ def chunks(l, n):
     """Yield successive n-sized chunks from l."""
     for i in range(0, len(l), n):
         yield l[i:i + n]
-
-
-def handle_exceptions(func, with_debugger):
-    """Drops a user into an interactive debugger if func raises an error."""
-
-    @functools.wraps(func)
-    def wrapped(*args, **kwargs):
-        try:
-            func(*args, **kwargs)
-        except (BdbQuit, KeyboardInterrupt):
-            raise
-        except Exception as e:
-            logger.exception("Uncaught exception {}".format(e))
-            if with_debugger:
-                import pdb
-                import traceback
-                traceback.print_exc()
-                pdb.post_mortem()
-            else:
-                raise
-
-    return wrapped
 
 
 def parse_package_version(s: str) -> Tuple[str, str]:
